Added default values to BAG esri exports
The functionality to assign default values was moved to utils to allow both the CSV en SHP exporter to use this option. Some small changes made to BAG export definitions
