+++ src/gobexport/exporter/utils.py
@@ -1,3 +1,105 @@
+def split_field_reference(ref: str):
+    return ref.split('.') if '.' in ref else ref
+
+
+def evaluate_condition(entity: dict, condition: dict):
+    """Expects an entity and a condition as dict, which has as keys:
+
+    condition: for example 'isempty'
+    reference: the field reference to which to apply the condition
+    trueval: the value (field reference) to return if condition evaluates to true
+    falseval: you get it
+
+    negate:    optional. if true, negate the condition
+
+    :param condition:
+    :return:
+    """
+    assert all([k in condition for k in ['condition', 'reference', 'trueval']]), "Invalid condition definition"
+
+    condition_type = condition.get('condition')
+    reference = condition.get('reference')
+    trueval = condition.get('trueval')
+    falseval = condition.get('falseval')
+    negate = condition.get('negate', False)
+    condition_should_be = not negate
+
+    onfield_value = get_entity_value(entity, split_field_reference(reference))
+
+    if condition_type == 'isempty':
+        condition_result = not bool(onfield_value)
+    else:
+        raise NotImplementedError(f"Not implemented condition f{condition_type}")
+
+    if condition_result is condition_should_be:
+        result = trueval
+    else:
+        result = falseval
+
+    if result:
+        return get_entity_value(entity, result)
+
+
+def _evaluate_concat_action(entity: dict, action: dict):
+    assert 'fields' in action
+    return "".join([str(item) if item is not None else "" for
+                    item in [get_entity_value(entity, field) for field in action['fields']]])
+
+
+def _evaluate_literal_action(action: dict):
+    return action.get('value')
+
+
+def evaluate_action(entity: dict, action: dict):
+    if action.get('action') == 'concat':
+        return _evaluate_concat_action(entity, action)
+    elif action.get('action') == 'literal':
+        return _evaluate_literal_action(action)
+    else:
+        raise NotImplementedError()
+
+
+def _get_entity_value_dict_lookup_key(entity: dict, lookup_key: dict):
+    if lookup_key.get('action'):
+        return evaluate_action(entity, lookup_key)
+    elif lookup_key.get('condition'):
+        return evaluate_condition(entity, lookup_key)
+
+    raise NotImplementedError()
+
+
+def get_entity_value(entity, lookup_key):
+    """Get the value from the entity using a key or a list of keys
+
+    lookup_key can be of type dict, with either 'action' or 'condition' set.
+
+    An action overrides the remainder of this function and returns the value immediately.
+    A condition returns a new (or the same) lookup key/dict for further processing.
+
+    If lookup_key is of type list or string, we perform a lookup on entity.
+
+    :param entity: The API entity to get the value from
+    :param lookup_key: A attribute name or a list of attribute names
+    :return: the value of the entity's attribute or None
+    """
+    if not lookup_key:
+        return None
+
+    if isinstance(lookup_key, str):
+        lookup_key = split_field_reference(lookup_key)
+
+    if isinstance(lookup_key, dict):
+        return _get_entity_value_dict_lookup_key(entity, lookup_key)
+
+    assert isinstance(lookup_key, str) or isinstance(lookup_key, list)
+
+    value = nested_entity_get(entity, lookup_key) if isinstance(lookup_key, list) else entity.get(lookup_key)
+    # Return J or N when the value is a boolean
+    if isinstance(value, bool):
+        value = 'J' if value else 'N'
+    return value
+
+
 def nested_entity_get(entity, keys, default=None):
     """
     Tries to get a value from a nested entity by a list of keys to look into.
